+++ packages/ricos-editor/web/src/toolbars/TextFormattingToolbar.tsx
@@ -93,6 +93,10 @@ class TextFormattingToolbar extends Component<TextFormattingToolbarProps> {
       linkSettings: this.props.linkSettings,
       isMobile,
     };
+    const headingsData = {
+      ...this.props.plugins?.find(plugin => plugin.type === 'wix-rich-content-plugin-headings')
+        ?.config,
+    };
     const onInlineToolbarOpen = () => this.props.onInlineToolbarOpen?.(ToolbarType.FORMATTING);
     const onToolbarButtonClick = (name, value = undefined, pluginId = undefined) => {
       this.props.onToolbarButtonClick?.(name, ToolbarType.FORMATTING, value, pluginId);
@@ -107,6 +111,7 @@ class TextFormattingToolbar extends Component<TextFormattingToolbarProps> {
         buttons={filteredFormattingToolbarButtons}
         linkPanelData={linkPanelData}
         colorPickerData={colorPickerData}
+        headingsData={headingsData}
         onToolbarButtonClick={onToolbarButtonClick}
         experiments={experiments}
       />