+++ packages/toolbars-new/web/src/Toolbar/buttonsListCreator.tsx
@@ -34,6 +34,7 @@ export const createButtonsList = (
   t,
   linkPanelData,
   colorPickerData,
+  headingsData,
   experiments
 ) => {
   const buttonsList = [];
@@ -51,7 +52,7 @@ export const createButtonsList = (
     handleButtonOnClick(buttonsList, index, editorCommands, linkPanelData, experiments);
     handleButtonIsActive(buttonsList, index, editorCommands);
     handleButtonIsDisabled(buttonsList, index, editorCommands);
-    handleButtonModal(buttonsList, index, editorCommands, linkPanelData, t);
+    handleButtonModal(buttonsList, index, editorCommands, linkPanelData, headingsData, t);
     handleButtonOnSave(buttonsList, index, editorCommands);
     handleButtonOnCancel(buttonsList, index, editorCommands);
     handleButtonOnChange(buttonsList, index, editorCommands);
@@ -64,8 +65,15 @@ export const createButtonsList = (
     handleButtonLoadSelection(buttonsList, index, editorCommands);
     handleButtonColorPicker(buttonsList, index, editorCommands, colorPickerData);
     handleButtonText(buttonsList, index, editorCommands, t);
+    handleButtonIsInput(buttonsList, index);
   });
   return buttonsList;
+};
+
+const handleButtonIsInput = (buttonsList, index) => {
+  if (buttonsFullData[buttonsList[index].name].isInput) {
+    buttonsList[index].isInput = buttonsFullData[buttonsList[index].name].isInput;
+  }
 };
 
 const handleButtonText = (buttonsList, index, editorCommands: editorCommands, t) => {
@@ -137,9 +145,9 @@ const handleButtonOnDone = (buttonsList, index, editorCommands: editorCommands) 
 const handleButtonOnChange = (buttonsList, index, editorCommands: editorCommands) => {
   if (buttonsFullData[buttonsList[index].name].onChange) {
     const buttonName = buttonsList[index].name;
-    if (buttonName === 'LINE_SPACING') {
-      buttonsList[index].onChange = type => {
-        updateSpacing(type, editorCommands, buttonName);
+    if (['LINE_SPACING', 'FONT_SIZE'].includes(buttonName)) {
+      buttonsList[index].onChange = value => {
+        updateDynamicStyles(value, editorCommands, buttonName);
       };
     }
   }
@@ -202,16 +210,17 @@ const handleTitleButton = (buttonsList, index, editorCommands: editorCommands) =
 const handleButtonOnSave = (buttonsList, index, editorCommands: editorCommands) => {
   if (buttonsFullData[buttonsList[index].name].onSave) {
     const buttonName = buttonsList[index].name;
-
     if (Object.keys(textBlockButtons).includes(buttonName)) {
-      buttonsList[index].onSave = type => editorCommands.setBlockType(type);
+      buttonsList[index].onSave = type => {
+        editorCommands.setBlockType(type);
+      };
     } else if (buttonName === 'Alignment') {
       buttonsList[index].onSave = type => editorCommands.setTextAlignment(type);
     } else if (Object.keys(decorationButtons).includes(buttonName)) {
-      buttonsList[index].onSave = type => {
-        if (buttonName === 'LINE_SPACING') {
-          if (type) {
-            updateSpacing(type, editorCommands, buttonName);
+      buttonsList[index].onSave = value => {
+        if (['LINE_SPACING', 'FONT_SIZE'].includes(buttonName)) {
+          if (value) {
+            updateDynamicStyles(value, editorCommands, buttonName);
             setTimeout(() => editorCommands.loadSelectionState());
           } else {
             editorCommands.loadEditorState();
@@ -227,6 +236,7 @@ const handleButtonModal = (
   index,
   editorCommands: editorCommands,
   linkPanelData,
+  headingsData,
   t
 ) => {
   const buttonName = buttonsList[index].name;
@@ -263,6 +273,10 @@ const handleButtonModal = (
             isMobileModalFullscreen
           />
         );
+    } else if (buttonName === 'FONT_SIZE') {
+      const Modal = buttonsFullData[buttonName].modal;
+      buttonsList[index].modal = props =>
+        Modal && <Modal {...props} currentSelect={getFontSize(editorCommands)} t={t} />;
     }
   }
 };
@@ -355,12 +369,20 @@ const handleButtonArrow = (buttonsList, index) => {
   }
 };
 
+const getFontSize = (editorCommands: editorCommands) => {
+  const fontSize = editorCommands.getFontSize() || '';
+  const pxRegex = new RegExp('[0-9]+[px]');
+  return pxRegex.exec(fontSize) ? fontSize.split('p')[0] : '';
+};
+
 const handleButtonLabel = (buttonsList, index, editorCommands: editorCommands, t) => {
   const buttonName = buttonsList[index].name;
   if (buttonsFullData[buttonName].label) {
     buttonsList[index].getLabel = () => buttonsFullData[buttonName].label;
     if (buttonName === 'HEADINGS') {
       buttonsList[index].getLabel = () => translateHeading(getCurrentHeading(editorCommands), t);
+    } else if (buttonName === 'FONT_SIZE') {
+      buttonsList[index].getLabel = () => getFontSize(editorCommands);
     }
   }
 };
@@ -490,9 +512,12 @@ const getCurrentHeading = (editorCommands: editorCommands) => {
   return currentHeading;
 };
 
-const updateSpacing = (type, editorCommands: editorCommands, buttonName) => {
-  const dynamicStyles = type;
-  editorCommands.insertDecoration(decorationButtons[buttonName], { dynamicStyles });
+const updateDynamicStyles = (value, editorCommands: editorCommands, buttonName) => {
+  const data =
+    buttonName === 'FONT_SIZE'
+      ? { fontSize: value < 10 ? 10 : value > 96 ? 96 : value }
+      : { dynamicStyles: value };
+  editorCommands.insertDecoration(decorationButtons[buttonName], { ...data });
 };
 
 const goToLink = (event, linkData, linkPanelData, experiments) => {