+++ packages/toolbars-new/web/src/Toolbar/buttonsListCreatorConsts.tsx
@@ -8,6 +8,7 @@ import {
   BULLET_LIST_TYPE,
   RICOS_INDENT_TYPE,
   RICOS_LINE_SPACING_TYPE,
+  RICOS_FONT_SIZE_TYPE,
   RICOS_LINK_TYPE,
   InlineStyle,
   DecorationsDataMap,
@@ -48,6 +49,7 @@ import LinkModal from '../modals/link/LinkComponents/LinkModal';
 import AlignmentPanel from '../modals/alignment/AlignmentPanel';
 import HeadingsPanel from '../modals/heading/HeadingsPanel';
 import LineSpacingPanel from '../modals/line-spacing/LineSpacingPanel';
+import FontSizePanel from '../modals/fontSize/FontSizePanel';
 
 export const HEADING_TYPE_TO_ELEMENT = Object.freeze({
   'header-one': 'H1',
@@ -133,6 +135,7 @@ type buttonsFullDataType = {
   unstyled?: { icon: any; action: string };
   'header-two'?: { icon: any; action: string };
   'header-three'?: { icon: any; action: string };
+  isInput?: boolean;
 };
 
 export const buttonsFullData: Record<string, buttonsFullDataType> = {
@@ -164,6 +167,20 @@ export const buttonsFullData: Record<string, buttonsFullDataType> = {
     type: 'modal',
     modal: props => <HeadingsPanel {...props} translateHeading={translateHeading} />,
     onSave: 'HEADINGS',
+    saveSelection: true,
+    loadSelection: true,
+  },
+  FONT_SIZE: {
+    icon: () => null,
+    dataHook: 'customFontSizeButton',
+    tooltip: 'FormattingToolbar_CustomFontSizeButton_Tooltip',
+    arrow: true,
+    type: 'modal',
+    modal: props => <FontSizePanel {...props} />,
+    onSave: 'FONT_SIZE',
+    onChange: 'FONT_SIZE',
+    label: 'FONT_SIZE',
+    isInput: true,
     saveSelection: true,
     loadSelection: true,
   },
@@ -430,6 +447,20 @@ export const decorationButtons: Record<string, keyof DecorationsDataMap> = {
   LINK: RICOS_LINK_TYPE,
   removeLink: RICOS_LINK_TYPE,
   editLink: RICOS_LINK_TYPE,
+  FONT_SIZE: RICOS_FONT_SIZE_TYPE,
+};
+
+export const deleteDecorationButtons: Record<
+  string,
+  | typeof RICOS_FONT_SIZE_TYPE
+  | typeof RICOS_TEXT_COLOR_TYPE
+  | typeof RICOS_TEXT_HIGHLIGHT_TYPE
+  | typeof RICOS_LINK_TYPE
+> = {
+  FONT_SIZE: RICOS_FONT_SIZE_TYPE,
+  TEXT_COLOR: RICOS_TEXT_COLOR_TYPE,
+  TEXT_HIGHLIGHT: RICOS_TEXT_HIGHLIGHT_TYPE,
+  LINK: RICOS_LINK_TYPE,
 };
 
 export const setTextAlignment: Record<string, TextAlignment> = {