+++ packages/toolbars-new/web/src/modals/line-spacing/LineSpacingPanel.jsx
@@ -64,7 +64,7 @@ class LineSpacingPanel extends Component {
           options: lineHeights,
           onChange: onSaveLineHeight,
           panelHeader,
-          showCustomPanel,
+          customPanelOptions: { inline: false, onOpen: showCustomPanel },
           t,
         }}
       />